Refactor xtask: modularize code by splitting main.rs into config.rs, commands.rs, and tasks.rs
